+++ go-functions/tor_lab_xxns/tor_lab_xxns.go
@@ -35,16 +35,16 @@ func Parse(w http.ResponseWriter, r *http.Request) {
 }
 
 type torLabxxnsStruct struct {
-	ID           uint8   `json:"id"`
-	BatteryLevel float32 `json:"battery_level"`
-	OpenState    bool    `json:"open_state"`
-	InternalData string  `json:"internal_data"`
+	ID           uint8  `json:"id"`
+	BatteryLevel uint8  `json:"battery_level"`
+	OpenState    bool   `json:"open_state"`
+	InternalData string `json:"internal_data"`
 }
 
 func (t *torLabxxnsStruct) parse(payload []byte) {
 	t.ID = uint8(payload[0])
 	// battery level expressed in 1/254 %
-	t.BatteryLevel = float32(payload[1]) / 254.0 * 100
+	t.BatteryLevel = uint8(float32(payload[1]) / 254.0 * 100)
 	t.OpenState = payload[2]&0x80 > 0
 	for _, b := range payload[3:] {
 		t.InternalData += fmt.Sprintf("%02x", b)
